Generate the vector embedding and then store it in the vector database and postgress.
The commit uses postgres with pgvector which helps in vector embedding
based similarity search rather than keyword based search for the
retrival part.(This forms the Javascript backend which is still in building stage). Also for converting the text to vector embeddings, a open
source model called all-MiniLM-L6-v2 from sentence transformer is
leveraged. This forms the python based backend.