Add search feature in the frontend
This feature allows addition of new frontend feature in the frontend. It
lets user search for the data, and it returns the relevant data in the
form of json.
In addition to the search feature, this commit replaces the previous
postgres+ pggvector database with the qdrant vector DB as it allows
metadata filtering much efficiently.